+++ Questionable/Windows/JournalComponents/QuestJournalComponent.cs
@@ -32,12 +32,8 @@ internal sealed class QuestJournalComponent
     IDalamudPluginInterface pluginInterface,
     QuestJournalUtils questJournalUtils,
     QuestValidator questValidator,
-    MovementController movementController,
-    AetheryteFunctions aetheryteFunctions,
-    AetheryteData aetheryteData,
     QuestController questController,
-    RedoUtil redoUtil,
-    IGameGui gameGui)
+    RedoUtil redoUtil)
 {
     private readonly Dictionary<JournalData.Category, JournalCounts> _categoryCounts = [];
     private readonly Dictionary<JournalData.Genre, JournalCounts> _genreCounts = [];
@@ -218,28 +214,7 @@ internal sealed class QuestJournalComponent
             questTooltipComponent.Draw(questInfo);
 
         if (ImGui.IsItemClicked())
-        {
-            var location = questInfo.IssuerLocation;
-            Svc.Log.Debug(location.ToString() ?? "SheetLevel()");
-            var mapLink = new MapLinkPayload(
-                location.Territory.RowId,
-                location.Map.RowId,
-                location.Game.X,
-                location.Game.Z
-            );
-            bool openedMap = gameGui.OpenMapWithMapLink(mapLink);
-            if (location.Territory.RowId.Equals(Svc.ClientState.TerritoryType))
-                movementController.NavigateTo(EMovementType.None, questInfo.IssuerDataId, location.Position, new()
-                {
-                    Fly = GameFunctions.IsFlyingUnlocked(location.Territory.RowId) ? true : false,
-                    Sprint = true,
-                    StopDistance = 20f,
-                    VerticalStopDistance = 5f,
-                });
-            else
-                if (aetheryteData.NearestAetheryteTo(location.Territory.RowId, location.Position) is { } aetheryte)
-                    aetheryteFunctions.TeleportAetheryte(aetheryte);
-        }
+            questJournalUtils.MoveToQuestLocation(questInfo);
 
         questJournalUtils.ShowContextMenu(questInfo, quest, nameof(QuestJournalComponent));
 
@@ -302,7 +277,7 @@ internal sealed class QuestJournalComponent
 
         ImGui.TableNextColumn();
         (Vector4 color, FontAwesomeIcon icon, string text) = uiUtils.GetQuestStyle(questInfo.QuestId);
-        uiUtils.ChecklistItem(text, color, icon);
+        uiUtils.ChecklistItem(text.Split(',',1)[0], color, icon);
     }
 
     internal static void DrawCount(int count, int total)