+++ Questionable/Windows/UiUtils.cs
@@ -16,7 +16,14 @@ internal sealed class UiUtils(QuestFunctions questFunctions, IDalamudPluginInter
 {
     public (Vector4 Color, FontAwesomeIcon Icon, string Status) GetQuestStyle(ElementId elementId)
     {
+        string lockedReason = string.Empty;
         HashSet<IQuestInfo>? prereqValue = null;
+        if (questFunctions.IsQuestLocked(elementId) is (bool isLocked, string[] reasons) && isLocked)
+            lockedReason = reasons[0];
+        else if (questFunctions.prereqCache.TryGetValue(elementId.Value, out prereqValue) &&
+                prereqValue.Any(q => questFunctions.IsQuestLocked(q.QuestId) is (bool qIsLocked, string[] reasons) && qIsLocked))
+            lockedReason = _L("Prev quest");
+
         if (questFunctions.IsQuestAccepted(elementId))
             return (ImGuiColors.DalamudYellow, FontAwesomeIcon.PersonWalkingArrowRight, _L("已接取"));
         else if (elementId is QuestId questId && questFunctions.IsDailyAlliedSocietyQuestAndAvailableToday(questId))
@@ -32,10 +39,8 @@ internal sealed class UiUtils(QuestFunctions questFunctions, IDalamudPluginInter
             return (ImGuiColors.ParsedGreen, FontAwesomeIcon.Check, _L("Complete"));
         else if (questFunctions.IsQuestUnobtainable(elementId))
             return (ImGuiColors.DalamudGrey, FontAwesomeIcon.Minus, _L("Unobtainable"));
-        else if (questFunctions.IsQuestLocked(elementId) is (bool isLocked, var _) && isLocked &&
-                questFunctions.prereqCache.TryGetValue(elementId.Value, out prereqValue) &&
-                prereqValue.Any(q => questFunctions.IsQuestLocked(q.QuestId) is (bool qIsLocked, string[] reasons) && qIsLocked))
-            return (ImGuiColors.DalamudRed, FontAwesomeIcon.Times, _L("Locked"));
+        else if (!string.IsNullOrEmpty(lockedReason))
+            return (ImGuiColors.DalamudRed, FontAwesomeIcon.Times, $"{_L("Locked")}: {lockedReason}");
         else if (prereqValue == null)
             return (ImGuiColors.TankBlue, FontAwesomeIcon.QuestionCircle, _L("Available(?)"));
         else